separate specialized vec implementations into files
