+++ SourceCode/include/image/image_factory.hpp
@@ -2,6 +2,7 @@
 #include <cstdint>
 #include <vector>
 #include "renderer/color.hpp"
+#include "util/vec2.hpp"
 
 std::vector<Color> create_random_color_rectangles_image(const cm::Vec2u resolution, const uint32_t rectangle_count_x = 4, const uint32_t rectangle_count_y = 4);
 std::vector<Color> create_fix_color_rectangles_image(const cm::Vec2u resolution, const uint32_t rectangle_count_x = 4, const uint32_t rectangle_count_y = 4);
