+++ SourceCode/include/object/material.hpp
@@ -5,7 +5,7 @@
 #include "object/texture.hpp"
 #include "renderer/hit_info.hpp"
 #include "renderer/ray.hpp"
-#include "util/vec.hpp"
+#include "util/vec3.hpp"
 
 enum class MaterialType
 {
