+++ SourceCode/include/object/texture.hpp
@@ -2,7 +2,8 @@
 
 #include <vector>
 #include <memory>
-#include "util/vec.hpp"
+#include "util/vec2.hpp"
+#include "util/vec3.hpp"
 
 enum class TextureType
 {
