+++ SourceCode/include/object/vertex.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
-#include "util/vec.hpp"
+#include "util/vec2.hpp"
+#include "util/vec3.hpp"
 
 struct Vertex
 {
